# [Perf] Inspector refactoring and optimization