--- a/pkg/inspector/inspector_benchmark_test.go
+++ b/pkg/inspector/inspector_benchmark_test.go
@@ -22,9 +22,28 @@ import (
 	"github.com/conduitio/conduit/pkg/record"
 )
 
+func BenchmarkInspector_NoSession_Send(b *testing.B) {
+	ins := New(log.Nop(), 10)
+
+	for i := 0; i < b.N; i++ {
+		ins.Send(context.Background(), record.Record{Position: record.Position("test-pos")})
+	}
+}
+
 func BenchmarkInspector_SingleSession_Send(b *testing.B) {
 	ins := New(log.Nop(), 10)
 	ins.NewSession(context.Background(), "test-id")
+
+	for i := 0; i < b.N; i++ {
+		ins.Send(context.Background(), record.Record{Position: record.Position("test-pos")})
+	}
+}
+
+func BenchmarkInspector_10Sessions_Send(b *testing.B) {
+	ins := New(log.Nop(), 10)
+	for i := 0; i < 10; i++ {
+		ins.NewSession(context.Background(), "test-id")
+	}
 
 	for i := 0; i < b.N; i++ {
 		ins.Send(context.Background(), record.Record{Position: record.Position("test-pos")})