--- a/pkg/inspector/inspector_test.go
+++ b/pkg/inspector/inspector_test.go
@@ -123,7 +123,7 @@ func TestInspector_Send_SlowConsumer(t *testing.T) {
 	s := underTest.NewSession(context.Background(), "test-id")
 
 	for i := 0; i < bufferSize+1; i++ {
-		s.send(
+		underTest.Send(
 			context.Background(),
 			record.Record{
 				Position: record.Position(fmt.Sprintf("test-pos-%v", i)),