### Description

The inspector was allocating quite a lot of data even if no sessions were open. Now it's optimized for the hot path when there are no open sessions.

```
goos: darwin
goarch: amd64
pkg: github.com/conduitio/conduit/pkg/inspector
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
                                │  bench.old  │              bench.new               │
                                │   sec/op    │    sec/op     vs base                │
Inspector_NoSession_Send-16       31.96n ± 2%    14.70n ± 1%  -54.01% (p=0.000 n=10)
Inspector_SingleSession_Send-16   92.88n ± 1%   122.15n ± 2%  +31.52% (p=0.000 n=10)
Inspector_10Sessions_Send-16      447.6n ± 2%    447.1n ± 2%        ~ (p=0.895 n=10)
geomean                           109.9n         92.94n       -15.46%
```

Note that the single session benchmark is slower now because of the use of `Record.Clone()`, although that usage is now correct, since we previously did not clone all fields. Note that cloning of records was sped up in #1247.

### Quick checks:

- [X] I have followed the [Code Guidelines](https://github.com/ConduitIO/conduit/blob/main/docs/code_guidelines.md).
- [X] There is no other [pull request](https://github.com/ConduitIO/conduit/pulls) for the same update/change.
- [X] I have written unit tests. (existing tests)
- [X] I have made sure that the PR is of reasonable size and can be easily reviewed.